+++ tests/test_bulk.py
@@ -11,18 +11,47 @@ class TestBulk(test.TruncationTestCase):
         await UniqueName.bulk_create([UniqueName() for _ in range(1000)])
         all_ = await UniqueName.all().values("id", "name")
         inc = all_[0]["id"]
-        self.assertListSortEqual(all_, [{"id": val + inc, "name": None} for val in range(1000)])
+        self.assertListSortEqual(
+            all_, [{"id": val + inc, "name": None} for val in range(1000)], sorted_key="id"
+        )
+
+    async def test_bulk_create_update_fields(self):
+        await UniqueName.bulk_create([UniqueName(name="name")])
+        await UniqueName.bulk_create(
+            [UniqueName(name="name", optional="optional")],
+            update_fields=["optional"],
+            on_conflict=["name"],
+        )
+        all_ = await UniqueName.all().values("name", "optional")
+        self.assertListSortEqual(all_, [{"name": "name", "optional": "optional"}])
+
+    async def test_bulk_create_more_that_one_update_fields(self):
+        await UniqueName.bulk_create([UniqueName(name="name")])
+        await UniqueName.bulk_create(
+            [UniqueName(name="name", optional="optional", other_optional="other_optional")],
+            update_fields=["optional", "other_optional"],
+            on_conflict=["name"],
+        )
+        all_ = await UniqueName.all().values("name", "optional", "other_optional")
+        self.assertListSortEqual(
+            all_, [{"name": "name", "optional": "optional", "other_optional": "other_optional"}]
+        )
 
     async def test_bulk_create_with_batch_size(self):
-        await UniqueName.bulk_create([UniqueName() for _ in range(1000)], batch_size=100)
+        await UniqueName.bulk_create(
+            [UniqueName(id=id_ + 1) for id_ in range(1000)], batch_size=100
+        )
         all_ = await UniqueName.all().values("id", "name")
-        inc = all_[0]["id"]
-        self.assertListSortEqual(all_, [{"id": val + inc, "name": None} for val in range(1000)])
+        self.assertListSortEqual(
+            all_, [{"id": val + 1, "name": None} for val in range(1000)], sorted_key="id"
+        )
 
     async def test_bulk_create_with_specified(self):
         await UniqueName.bulk_create([UniqueName(id=id_) for id_ in range(1000, 2000)])
         all_ = await UniqueName.all().values("id", "name")
-        self.assertListSortEqual(all_, [{"id": id_, "name": None} for id_ in range(1000, 2000)])
+        self.assertListSortEqual(
+            all_, [{"id": id_, "name": None} for id_ in range(1000, 2000)], sorted_key="id"
+        )
 
     async def test_bulk_create_mix_specified(self):
         await UniqueName.bulk_create(
@@ -34,11 +63,11 @@ class TestBulk(test.TruncationTestCase):
         self.assertEqual(len(all_), 2000)
 
         self.assertListSortEqual(
-            all_[:1000], [{"id": id_, "name": None} for id_ in range(10000, 11000)]
+            all_[:1000], [{"id": id_, "name": None} for id_ in range(10000, 11000)], sorted_key="id"
         )
         inc = all_[1000]["id"]
         self.assertListSortEqual(
-            all_[1000:], [{"id": val + inc, "name": None} for val in range(1000)]
+            all_[1000:], [{"id": val + inc, "name": None} for val in range(1000)], sorted_key="id"
         )
 
     async def test_bulk_create_uuidpk(self):
@@ -90,3 +119,10 @@ class TestBulk(test.TruncationTestCase):
         with self.assertRaises(IntegrityError):
             async with in_transaction():
                 await UUIDPkModel.bulk_create([UUIDPkModel(id=val) for _ in range(10)])
+
+    async def test_bulk_create_ignore_conflicts(self):
+        name1 = UniqueName(name="name1")
+        name2 = UniqueName(name="name2")
+        await UniqueName.bulk_create([name1, name2], ignore_conflicts=True)
+        with self.assertRaises(IntegrityError):
+            await UniqueName.bulk_create([name1, name2])