+++ tortoise/backends/base/client.py
@@ -226,7 +226,7 @@ class TransactionContext:
         self.lock = getattr(connection, "_trxlock", None)
 
     async def __aenter__(self):
-        await self.lock.acquire()
+        await self.lock.acquire()  # type:ignore
         current_transaction = current_transaction_map[self.connection_name]
         self.token = current_transaction.set(self.connection)
         await self.connection.start()
@@ -241,7 +241,7 @@ class TransactionContext:
             else:
                 await self.connection.commit()
         current_transaction_map[self.connection_name].reset(self.token)
-        self.lock.release()
+        self.lock.release()  # type:ignore
 
 
 class TransactionContextPooled(TransactionContext):
@@ -281,11 +281,11 @@ class NestedTransactionContext(TransactionContext):
 
 class NestedTransactionPooledContext(TransactionContext):
     async def __aenter__(self):
-        await self.lock.acquire()
+        await self.lock.acquire()  # type:ignore
         return self.connection
 
     async def __aexit__(self, exc_type: Any, exc_val: Any, exc_tb: Any) -> None:
-        self.lock.release()
+        self.lock.release()  # type:ignore
         if not self.connection._finalized:
             if exc_type:
                 # Can't rollback a transaction that already failed.