+++ tortoise/backends/base/executor.py
@@ -20,6 +20,7 @@ from typing import (
 )
 
 from pypika import JoinType, Parameter, Query, Table
+from pypika.queries import QueryBuilder
 from pypika.terms import ArithmeticExpression, Function
 
 from tortoise.exceptions import OperationalError
@@ -69,15 +70,15 @@ class BaseExecutor:
         key = f"{self.db.connection_name}:{self.model._meta.db_table}"
         if key not in EXECUTOR_CACHE:
             self.regular_columns, columns = self._prepare_insert_columns()
-            self.insert_query = self._prepare_insert_statement(columns)
+            self.insert_query = str(self._prepare_insert_statement(columns))
             self.regular_columns_all = self.regular_columns
             self.insert_query_all = self.insert_query
             if self.model._meta.generated_db_fields:
                 self.regular_columns_all, columns_all = self._prepare_insert_columns(
                     include_generated=True
                 )
-                self.insert_query_all = self._prepare_insert_statement(
-                    columns_all, has_generated=False
+                self.insert_query_all = str(
+                    self._prepare_insert_statement(columns_all, has_generated=False)
                 )
 
             self.column_map: Dict[str, Callable[[Any, Any], Any]] = {}
@@ -195,15 +196,20 @@ class BaseExecutor:
             return cls.TO_DB_OVERRIDE[field_object.__class__](field_object, attr, instance)
         return field_object.to_db_value(attr, instance)
 
-    def _prepare_insert_statement(self, columns: Sequence[str], has_generated: bool = True) -> str:
+    def _prepare_insert_statement(
+        self, columns: Sequence[str], has_generated: bool = True, ignore_conflicts: bool = False
+    ) -> QueryBuilder:
         # Insert should implement returning new id to saved object
         # Each db has it's own methods for it, so each implementation should
         # go to descendant executors
-        return str(
+        query = (
             self.db.query_class.into(self.model._meta.basetable)
             .columns(*columns)
             .insert(*[self.parameter(i) for i in range(len(columns))])
         )
+        if ignore_conflicts:
+            query = query.do_nothing()
+        return query
 
     async def _process_insert_result(self, instance: "Model", results: Any) -> None:
         raise NotImplementedError()  # pragma: nocoverage